+++ inference/core/entities/responses/doctr.py
@@ -1,0 +1,16 @@
+from pydantic import BaseModel, Field
+
+
+class DoctrOCRInferenceResponse(BaseModel):
+    """
+    DocTR Inference response.
+
+    Attributes:
+        result (str): The result from OCR.
+        time: The time in seconds it took to produce the segmentation including preprocessing.
+    """
+
+    result: str = Field(description="The result from OCR.")
+    time: float = Field(
+        description="The time in seconds it took to produce the segmentation including preprocessing."
+    )
