+++ inference/core/interfaces/http/http_api.py
@@ -15,6 +15,7 @@ from inference.core.entities.requests.clip import (
     ClipImageEmbeddingRequest,
     ClipTextEmbeddingRequest,
 )
+from inference.core.entities.requests.doctr import DoctrOCRInferenceRequest
 from inference.core.entities.requests.gaze import GazeDetectionInferenceRequest
 from inference.core.entities.requests.inference import (
     ClassificationInferenceRequest,
@@ -35,9 +36,11 @@ from inference.core.entities.responses.clip import (
     ClipCompareResponse,
     ClipEmbeddingResponse,
 )
+from inference.core.entities.responses.doctr import DoctrOCRInferenceResponse
 from inference.core.entities.responses.gaze import GazeDetectionInferenceResponse
 from inference.core.entities.responses.inference import (
     ClassificationInferenceResponse,
+    DoctrOCRInferenceResponse,
     InferenceResponse,
     InstanceSegmentationInferenceResponse,
     MultiLabelClassificationInferenceResponse,
@@ -54,6 +57,7 @@ from inference.core.entities.responses.server_state import (
 from inference.core.env import (
     ALLOW_ORIGINS,
     CORE_MODEL_CLIP_ENABLED,
+    CORE_MODEL_DOCTR_ENABLED,
     CORE_MODEL_GAZE_ENABLED,
     CORE_MODEL_SAM_ENABLED,
     CORE_MODELS_ENABLED,
@@ -63,7 +67,6 @@ from inference.core.env import (
     METRICS_ENABLED,
     PROFILE,
     ROBOFLOW_SERVICE_SECRET,
-    CORE_MODEL_DOCTR_ENABLED
 )
 from inference.core.exceptions import (
     ContentTypeInvalid,
@@ -614,15 +617,16 @@ class HttpInterface(BaseInterface):
                     return response
 
             if CORE_MODEL_DOCTR_ENABLED:
+
                 @app.post(
                     "/doctr/ocr",
-                    response_model=M.DoctrOCRInferenceResponse,
+                    response_model=DoctrOCRInferenceResponse,
                     summary="DocTR OCR response",
                     description="Run the DocTR OCR model to retrieve text in an image.",
                 )
                 @with_route_exceptions
                 async def doctr_retrieve_text(
-                    inference_request: M.DoctrOCRInferenceRequest,
+                    inference_request: DoctrOCRInferenceRequest,
                     api_key: Optional[str] = Query(
                         None,
                         description="Roboflow API Key that will be passed to the model during initialization for artifact retrieval",
