+++ inference/models/doctr/__init__.py
@@ -1,2 +1,1 @@
 from inference.models.doctr.doctr_model import DocTR
-
